god_hands_solver: stop complaining about the missing .depend
GNU make emitted a warning at first compilation because the provided Makefile
tries to include a .depend file which is not delivered in the repository.

This file is created at the first compilation and therefore the warning
is not emitted afterward.

To stop complaining about this file, we replaced the 'include' directive
by the '-include' which simply ignores the files that do not exist.